Woo: Append product variations to edit actions
When a product changes from variable to simple, or any time a products
variations attribute changes, the variation objects for that product
need to be adjusted. For the reducer to be able to add the appropriate
variation edits, it needs the current set of variations for reference.

This patch adds a middleware handler that appends the needed data to the
action before it reaches the reducer.