Woo: No longer create edit on product page loads
The previous code would create an edit on page load as a side effect of
setting the `currentlyEditingId`. This avoids that approach and only
sets the `currentlyEditingId`, which means that if an edit create or
update exists, there's a real data edit there.

This is requied for checking if it's okay for a user to navigate away
from the page.